chore(team): split mailbox thread helpers
